Parses URLs for images and Lottie animations
Introduces a URL parsing mechanism for image and Lottie animation URLs within the Ink story engine.

This change allows the application to handle different URL formats, including absolute URLs, application-relative URLs, and shorthand hostnames, ensuring that images and Lottie animations are loaded correctly regardless of the URL format used in the Ink story.

It also removes the redundant URL parsing logic from the Index.razor.cs file and centralizes it within the new UrlParser class for improved maintainability and code reuse.
